Add files via upload
Initial glitching software program. Uses G-Code to control the printer and has a GUI to set up the EMFI injection. I don't have the faulty cat yet so I can't set up the proper glitch detection and triggering yet but that will be done in a month or so when I have it.